Add --no-reboot flag

+++ setup.sh
@@ -33,8 +33,13 @@ if [[ "${BASH_SOURCE[0]}" == "${0}" ]]; then
   sudo rm -R /tmp/kiwisdr-conf-main
 
   echo "✅ The KiwiSDR configuration is complete."
-  echo "ℹ️ The KiwiSDR wil now reboot."
-  sudo reboot
+  if [ $1 == "--no-reboot" ]; then
+    echo "ℹ️ The --no-reboot flag was provided, so the KiwiSDR will not reboot now."
+    exit 0
+  else
+    echo "ℹ️ The KiwiSDR will now reboot."
+    sudo reboot
+  fi
 fi
